requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1 ; python_version >= '3.8'
attrs==24.2.0 ; python_version >= '3.7'
autobahn==24.4.2 ; python_version >= '3.9'
automat==24.8.1 ; python_version >= '3.8'
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
channels[daphne]==4.1.0
channels-redis==4.2.0
constantly==23.10.4 ; python_version >= '3.8'
cryptography==43.0.1 ; python_version >= '3.7'
daphne==4.1.2
django==4.2.16
django-environ==0.11.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2 ; python_version >= '3.8'
msgpack==1.1.0 ; python_version >= '3.8'
pyasn1==0.6.1 ; python_version >= '3.8'
pyasn1-modules==0.4.1 ; python_version >= '3.8'
pycparser==2.22 ; python_version >= '3.8'
pyopenssl==24.2.1
redis==5.0.8 ; python_version >= '3.7'
service-identity==24.1.0
setuptools==74.1.2 ; python_version >= '3.8'
sqlparse==0.5.1 ; python_version >= '3.8'
twisted[tls]==24.7.0 ; python_full_version >= '3.8.0'
txaio==23.1.1 ; python_version >= '3.7'
typing-extensions==4.12.2 ; python_version >= '3.8'
zope-interface==7.0.3 ; python_version >= '3.8'
